requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
gunicorn
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.0.0 
pyjwt==2.3.0
psycopg2==2.9.1